Constructed grid and added mouseover event listener to each grid box
The idea with the grid construction is first we create a loop to create
a set number of rows, and within each loop for each row we create another
loop to create a number of boxes within each row. Each row is given flex
property so that children boxes are appended horizontally. On creation of
each grid box, the grid box is given an addEventListener which changes the
background colour to brushColor when a pointer moves over the box.

The container width is set to a specific value, and we use container width
divided by number of grids to find a value for the width and height of each
box to make the grids fit nicely within the container space. I just realised
I should add box-sizing: border-box as well so that the borders on each grid
do not affect the size of the canvas.