Added way to clear grid, as well as choose new grid sizes
container.firstChild === null if no more children, so I set
if createGrid is passed with 0 as parameter, a loop is run that
removes all children of container div until there is no more
children.

Also added a reset button that first calls createGrid(0), then
asks user for prompt to choose a new size. Then, sizeOfGrid is set
to this new size, which is called in the new createGrid(userSize)
at the end of the reset-button event listener.

Weirdly, let userSize = Number.prompt("...") returns an error
saying number.prompt is not a function. So for now it is possible
to pass strings into the prompt...